requirements.txt
# ===============================
# Core Web Framework
# ===============================
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
python-multipart==0.0.9

# ===============================
# Deep Learning (CPU only)
# ===============================
torch==2.2.2
torchvision==0.17.2
pillow==10.2.0
numpy==1.26.4

# ===============================
# LLM + STT (Groq)
# ===============================
groq==0.9.0
python-dotenv==1.0.1

# ===============================
# Utilities
# ===============================
requests==2.31.0